+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 vNext
 - Added API to add and get EDP info. 
 - Added UI to show EDP metrics to homepage.
+- Added prod configuration for frontend and backend.